+++ controllers/crud_auth.go
@@ -2,7 +2,6 @@ package controllers
 
 import (
 	"fmt"
-	"html/template"
 
 	beego "github.com/beego/beego/v2/adapter"
 	"github.com/beego/beego/v2/client/orm"
@@ -33,8 +32,6 @@ func (c *loggedRouter) Prepare() {
 		c.Data["TabIcon"] = dico.Ctx.IconFile
 		c.Data["TabTitle"] = dico.Ctx.Title
 	}
-	// XSRF protection des formulaires
-	c.Data["xsrfdata"] = template.HTML(c.XSRFFormHTML())
 
 }