+++ controllers/crud_common.go
@@ -3,6 +3,7 @@ package controllers
 import (
 	"errors"
 	"fmt"
+	"html/template"
 	"reflect"
 	"regexp"
 	"sort"
@@ -508,6 +509,8 @@ func setContext(c beego.Controller, table string) {
 	config.Theme = beego.AppConfig.String("theme")
 	c.Data["Config"] = &config
 
+	// XSRF protection des formulaires
+	c.Data["xsrfdata"] = template.HTML(c.XSRFFormHTML())
 	// Title
 	c.Data["Title"] = config.Appname
 	c.Data["Portail"] = &dico.Ctx